+++ src/main/java/ru/mart/andersen/creditor/repository/ProductRepository.java
@@ -5,11 +5,12 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import ru.mart.andersen.creditor.model.Product;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
 public interface ProductRepository extends CrudRepository<Product, Long> {
 
     @Query("SELECT p from Product p WHERE p.minSum<=:sum AND p.maxSum>=:sum")
-    Optional<Product> findBySum(int sum);
+    Optional<Product> findBySum(BigDecimal sum);
 }
